+++ webpack/src/core/dom/getTypeOfRenderingTree.ts
@@ -1,0 +1,11 @@
+import { RenderingTree } from "../../interfaces/RenderingTree";
+
+export function getTypeOfRenderingTree(
+  renderingTree: RenderingTree
+): "text" | "element" | "use" {
+  return typeof renderingTree === "string"
+    ? "text"
+    : "type" in renderingTree
+    ? renderingTree.type
+    : "use";
+}
