+++ webpack/src/core/invoke.ts
@@ -1,6 +1,5 @@
 import { Action } from "../interfaces/Action";
-import { onAction } from "./nDom";
-
+import { onAction } from "./dom/onAction";
 export function invoke(action: Action) {
-  onAction(action);
+  onAction.get()(action);
 }
