+++ webpack/src/interfaces/Action.ts
@@ -1,3 +1,3 @@
 export interface Action {
-  type: "DateIndicate" | "DateIndicate";
+  type: "DateIndicate";
 }
