+++ webpack/src/interfaces/HtmlTagFunctions.ts
@@ -4,5 +4,5 @@ import { HtmlTagName } from '../constants/HtmlTag';
 type HtmlTagUnion = keyof typeof HtmlTagName;
 
 export type HtmlTagFunctions = {
-  [K in HtmlTagUnion]?: HtmlTagFunction;
+  [K in HtmlTagUnion]: HtmlTagFunction;
 };
